requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
kafka-python==2.0.2
sqlalchemy==2.0.30
psycopg[binary,pool]==3.1.18
asyncpg==0.29.0
langchain==0.2.1
langchain-community==0.2.1
langchain-ollama==0.1.0
pypdf==4.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
streamlit==1.35.0
ruff==0.4.4
mypy==1.10.0
pre-commit==3.7.1
python-dotenv==1.0.1
langchain-postgres==0.0.17